refactor(shop): split v1 shop controller by resource
