+++ src/Shop/Transport/Controller/Api/V1/Product/GetProductController.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Shop\Transport\Controller\Api\V1\Product;
+
+use App\Shop\Application\Service\ProductListService;
+use App\Shop\Domain\Entity\Product;
+use App\Shop\Infrastructure\Repository\ProductRepository;
+use OpenApi\Attributes as OA;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Attribute\AsController;
+use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\Authorization\Voter\AuthenticatedVoter;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+
+#[AsController]
+#[OA\Tag(name: 'Shop')]
+#[IsGranted(AuthenticatedVoter::IS_AUTHENTICATED_FULLY)]
+final readonly class GetProductController
+{
+    public function __construct(private ProductRepository $productRepository)
+    {
+    }
+
+    #[Route('/v1/shop/products/{id}', methods: [Request::METHOD_GET])]
+    public function __invoke(string $id): JsonResponse
+    {
+        $product = $this->productRepository->find($id);
+        if (!$product instanceof Product) {
+            return new JsonResponse(status: JsonResponse::HTTP_NOT_FOUND);
+        }
+
+        return new JsonResponse(ProductListService::serializeProduct($product));
+    }
+}
