+++ src/Shop/Transport/Controller/Api/V1/Product/ListProductsController.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Shop\Transport\Controller\Api\V1\Product;
+
+use App\Shop\Application\Service\ProductListService;
+use OpenApi\Attributes as OA;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Attribute\AsController;
+use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\Authorization\Voter\AuthenticatedVoter;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+
+#[AsController]
+#[OA\Tag(name: 'Shop')]
+#[IsGranted(AuthenticatedVoter::IS_AUTHENTICATED_FULLY)]
+final readonly class ListProductsController
+{
+    public function __construct(private ProductListService $productListService)
+    {
+    }
+
+    #[Route('/v1/shop/products', methods: [Request::METHOD_GET])]
+    public function __invoke(Request $request): JsonResponse
+    {
+        return new JsonResponse($this->productListService->getList($request));
+    }
+}
